back to stable

+++ model.py
@@ -28,7 +28,7 @@ augmented_images , augmented_measurements = [],[]
 for image , mesure in zip(imgs,labels):
     augmented_images.append(image)
     augmented_measurements.append(mesure)
-    image = cv2.resize(image, (66, 200)).astype(np.float32)
+    image = cv2.resize(image, (66, 200))
     augmented_images.append(cv2.flip(image,1))
     augmented_measurements.append(mesure*-1.0)
